+++ tests/embedders/hex2vec/constants.py
@@ -1,7 +1,7 @@
 """Constants for hex2vec tests."""
 
 ENCODER_SIZES = [32, 16]
-TRAINER_KWARGS = {"max_epochs": 1, "deterministic": True}
+TRAINER_KWARGS = {"max_epochs": 1}
 
 PREDEFINED_TEST_CASES = [
     {