+++ tests/regionalizers/test_voronoi_regionalizer.py
@@ -1,4 +1,4 @@
-"""Voronoi regionizer tests."""
+"""Voronoi regionalizer tests."""
 from typing import Any
 
 import geopandas as gpd
@@ -7,15 +7,15 @@ import pytest
 from shapely.geometry import Point, Polygon
 
 from srai.constants import GEOMETRY_COLUMN, WGS84_CRS
-from srai.regionizers import VoronoiRegionizer
-from srai.regionizers._spherical_voronoi import generate_voronoi_regions
+from srai.regionalizers import VoronoiRegionalizer
+from srai.regionalizers._spherical_voronoi import generate_voronoi_regions
 from srai.utils import merge_disjointed_gdf_geometries
 
 
 def test_empty_gdf_attribute_error(gdf_empty: gpd.GeoDataFrame) -> None:
     """Test checks if empty GeoDataFrames are disallowed."""
     with pytest.raises(AttributeError):
-        VoronoiRegionizer(seeds=gdf_empty)
+        VoronoiRegionalizer(seeds=gdf_empty)
 
 
 def test_no_crs_gdf_value_error(
@@ -23,7 +23,7 @@ def test_no_crs_gdf_value_error(
 ) -> None:
     """Test checks if GeoDataFrames without crs are disallowed."""
     with pytest.raises(ValueError):
-        vr = VoronoiRegionizer(seeds=gdf_earth_poles)
+        vr = VoronoiRegionalizer(seeds=gdf_earth_poles)
         vr.transform(gdf=gdf_no_crs)
 
 
@@ -35,7 +35,7 @@ def test_duplicate_seeds_value_error() -> None:
             index=[1, 2, 3, 4],
             crs=WGS84_CRS,
         )
-        VoronoiRegionizer(seeds=seeds_gdf)
+        VoronoiRegionalizer(seeds=seeds_gdf)
 
 
 def test_single_seed_region() -> None:
@@ -46,7 +46,7 @@ def test_single_seed_region() -> None:
             index=[1],
             crs=WGS84_CRS,
         )
-        VoronoiRegionizer(seeds=seeds_gdf)
+        VoronoiRegionalizer(seeds=seeds_gdf)
 
 
 def test_single_seed_algorithm_error() -> None:
@@ -59,7 +59,7 @@ def test_multiple_seeds_regions(
     gdf_earth_poles: gpd.GeoDataFrame, gdf_earth_bbox: gpd.GeoDataFrame, earth_bbox: Polygon
 ) -> None:
     """Test checks if regions are generated correctly."""
-    vr = VoronoiRegionizer(seeds=gdf_earth_poles)
+    vr = VoronoiRegionalizer(seeds=gdf_earth_poles)
     result_gdf = vr.transform(gdf=gdf_earth_bbox)
     assert len(result_gdf.index) == 6
     assert merge_disjointed_gdf_geometries(result_gdf).difference(earth_bbox).is_empty
@@ -82,7 +82,7 @@ def test_big_number_of_seeds_regions(gdf_earth_bbox: gpd.GeoDataFrame, earth_bbo
         crs=WGS84_CRS,
     )
 
-    vr = VoronoiRegionizer(seeds=random_points_gdf)
+    vr = VoronoiRegionalizer(seeds=random_points_gdf)
     result_gdf = vr.transform(gdf=gdf_earth_bbox)
     assert len(result_gdf.index) == number_of_points
     assert merge_disjointed_gdf_geometries(result_gdf).difference(earth_bbox).is_empty
@@ -102,7 +102,7 @@ def test_four_close_seed_region(gdf_earth_bbox: gpd.GeoDataFrame, earth_bbox: Po
         index=[1, 2, 3, 4],
         crs=WGS84_CRS,
     )
-    vr = VoronoiRegionizer(seeds=seeds_gdf)
+    vr = VoronoiRegionalizer(seeds=seeds_gdf)
     result_gdf = vr.transform(gdf=gdf_earth_bbox)
     assert len(result_gdf.index) == 4
     assert merge_disjointed_gdf_geometries(result_gdf).difference(earth_bbox).is_empty
@@ -110,7 +110,7 @@ def test_four_close_seed_region(gdf_earth_bbox: gpd.GeoDataFrame, earth_bbox: Po
 
 def test_default_parameter(gdf_earth_poles: gpd.GeoDataFrame, earth_bbox: Polygon) -> None:
     """Test checks if regions are generated correctly with a default mask."""
-    vr = VoronoiRegionizer(seeds=gdf_earth_poles)
+    vr = VoronoiRegionalizer(seeds=gdf_earth_poles)
     result_gdf = vr.transform(gdf=None)
     assert len(result_gdf.index) == 6
     assert merge_disjointed_gdf_geometries(result_gdf).difference(earth_bbox).is_empty
@@ -131,6 +131,6 @@ def test_clipping_parameter(
     """Test checks if regions are clipped correctly with a provided mask."""
     gdf: gpd.GeoDataFrame = request.getfixturevalue(gdf_fixture)
     gdf_earth_poles: gpd.GeoDataFrame = request.getfixturevalue("gdf_earth_poles")
-    vr = VoronoiRegionizer(seeds=gdf_earth_poles)
+    vr = VoronoiRegionalizer(seeds=gdf_earth_poles)
     result_gdf = vr.transform(gdf=gdf)
     assert merge_disjointed_gdf_geometries(result_gdf).difference(gdf.iloc[0].geometry).is_empty